refactor and add functionality
