+++ src/components/map/filters/categoryFilter.vue
